+++ src/commands/ExportCommand.ts
@@ -1,5 +1,3 @@
-// src/commands/ExportCommand.ts
-
 import * as vscode from 'vscode'
 import { ConfigManager } from '../utils/ConfigManager'
 import { FileProcessor } from '../utils/FileProcessor'
@@ -21,16 +19,15 @@ export class ExportCommand {
 
     const rootPath = workspaceFolder.uri.fsPath
     const configManager = new ConfigManager(rootPath)
-    const config = await configManager.getConfig()
     const fileProcessor = new FileProcessor(
       rootPath,
-      config,
+      configManager,
       this.outputChannel
     )
 
     const structureGenerator = new ProjectStructureGenerator(
       rootPath,
-      config,
+      configManager,
       fileProcessor
     )
 
@@ -42,6 +39,7 @@ export class ExportCommand {
       },
       async (progress, token) => {
         try {
+          const config = await configManager.getConfig()
           const description = await configManager.getDescription('main')
           let output: string[] = []
 
@@ -50,11 +48,6 @@ export class ExportCommand {
           }
 
           if (config.includeProjectStructure) {
-            const structureGenerator = new ProjectStructureGenerator(
-              rootPath,
-              config,
-              fileProcessor
-            )
             output.push(await structureGenerator.generate())
           }